TestTrail/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==41.0.1
Django==4.2.1
django-cors-headers==4.1.0
django-crispy-forms==2.0
django-environ==0.10.0
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.8
pdfminer.six==20221105
psycopg2==2.9.6
pycparser==2.21
PyPDF2==3.0.1
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
